[platform/cel] Move platform api to under device platform
