+++ device/celestica/x86_64-cel_e1031-r0/sonic_platform/platform.py
@@ -10,7 +10,7 @@
 
 try:
     from sonic_platform_base.platform_base import PlatformBase
-    from sonic_platform_api.chassis import Chassis
+    from sonic_platform.chassis import Chassis
 except ImportError as e:
     raise ImportError(str(e) + "- required module not found")
