--- platform/broadcom/sonic-platform-modules-cel/cel-platform-api/sonic_platform_api/chassis.py
@@ -1,148 +1,0 @@
-#!/usr/bin/env python
-
-#############################################################################
-# Celestica
-#
-# Module contains an implementation of SONiC Platform Base API and
-# provides the Chassis information which are available in the platform
-#
-#############################################################################
-
-import sys
-import re
-import os
-import subprocess
-import json
-
-try:
-    from sonic_platform_base.chassis_base import ChassisBase
-    from sonic_platform_api.fan import Fan
-except ImportError as e:
-    raise ImportError(str(e) + "- required module not found")
-
-
-class Chassis(ChassisBase):
-    """Platform-specific Chassis class"""
-
-    def __init__(self):
-        ChassisBase.__init__(self)
-        self.config_data = {}
-        self.config_db_path = "/etc/sonic/config_db.json"
-        self.platform = self.get_platform()
-        NUM_FAN = {
-            "x86_64-cel_e1031-r0": 3,
-            "x86_64-cel_seastone-r0": 5
-        }.get(self.platform, 0)
-        for index in range(0, NUM_FAN):
-            fan = Fan(index)
-            self._fan_list.append(fan)
-        self.bios_version_path = "/sys/class/dmi/id/bios_version"
-        self.smc_cpld_e1031_path = "/sys/devices/platform/e1031.smc/version"
-        self.mmc_cpld_e1031_path = "/sys/devices/platform/e1031.smc/getreg"
-        self.cpld_dx010_path = "/sys/devices/platform/dx010_cpld/getreg"
-
-    def __get_register_value(self, path, register):
-        cmd = "echo {1} > {0}; cat {0}".format(path, register)
-        p = subprocess.Popen(
-            cmd, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-        raw_data, err = p.communicate()
-        if err is not '':
-            return 'None'
-        else:
-            return raw_data.strip()
-
-    def __read_config_db(self):
-        try:
-            with open(self.config_db_path, 'r') as fd:
-                data = json.load(fd)
-                return data
-        except IOError:
-            raise IOError("Unable to open config_db file !")
-
-    def get_platform(self):
-        try:
-            self.config_data = self.__read_config_db()
-            platform = self.config_data["DEVICE_METADATA"]["localhost"]["platform"]
-            return platform
-        except KeyError:
-            raise KeyError("Platform name not found")
-
-    def get_base_mac(self):
-        """
-        Retrieves the base MAC address for the chassis
-        Returns:
-            A string containing the MAC address in the format
-            'XX:XX:XX:XX:XX:XX'
-        """
-        try:
-            self.config_data = self.__read_config_db()
-            base_mac = self.config_data["DEVICE_METADATA"]["localhost"]["mac"]
-            return str(base_mac)
-        except KeyError:
-            raise KeyError("Base MAC not found")
-
-    def get_component_versions(self):
-        """
-        Retrieves platform-specific hardware/firmware versions for chassis
-        componenets such as BIOS, CPLD, FPGA, etc.
-        Returns:
-            A string containing platform-specific component versions
-        """
-
-        component_versions = dict()
-
-        # Get BIOS version
-        try:
-            with open(self.bios_version_path, 'r') as fd:
-                bios_version = fd.read()
-        except IOError:
-            raise IOError("Unable to open version file !")
-
-        # Get CPLD version
-        if self.platform == "x86_64-cel_seastone-r0":
-
-            CPLD_1 = self.__get_register_value(self.cpld_dx010_path, '0x100')
-            CPLD_2 = self.__get_register_value(self.cpld_dx010_path, '0x200')
-            CPLD_3 = self.__get_register_value(self.cpld_dx010_path, '0x280')
-            CPLD_4 = self.__get_register_value(self.cpld_dx010_path, '0x300')
-            CPLD_5 = self.__get_register_value(self.cpld_dx010_path, '0x380')
-
-            CPLD_1 = 'None' if CPLD_1 is 'None' else "{}.{}".format(
-                int(CPLD_1[2], 16), int(CPLD_1[3], 16))
-            CPLD_2 = 'None' if CPLD_2 is 'None' else "{}.{}".format(
-                int(CPLD_2[2], 16), int(CPLD_2[3], 16))
-            CPLD_3 = 'None' if CPLD_3 is 'None' else "{}.{}".format(
-                int(CPLD_3[2], 16), int(CPLD_3[3], 16))
-            CPLD_4 = 'None' if CPLD_4 is 'None' else "{}.{}".format(
-                int(CPLD_4[2], 16), int(CPLD_4[3], 16))
-            CPLD_5 = 'None' if CPLD_5 is 'None' else "{}.{}".format(
-                int(CPLD_5[2], 16), int(CPLD_5[3], 16))
-
-            cpld_version = dict()
-            cpld_version["CPLD1"] = CPLD_1
-            cpld_version["CPLD2"] = CPLD_2
-            cpld_version["CPLD3"] = CPLD_3
-            cpld_version["CPLD4"] = CPLD_4
-            cpld_version["CPLD5"] = CPLD_5
-            component_versions["CPLD"] = cpld_version
-
-        elif self.platform == "x86_64-cel_e1031-r0":
-            cpld_version = dict()
-
-            with open(self.smc_cpld_e1031_path, 'r') as fd:
-                smc_cpld_version = fd.read()
-            smc_cpld_version = 'None' if smc_cpld_version is 'None' else "{}.{}".format(
-                int(smc_cpld_version[2], 16), int(smc_cpld_version[3], 16))
-
-            mmc_cpld_version = self.__get_register_value(
-                self.mmc_cpld_e1031_path, '0x100')
-            mmc_cpld_version = 'None' if mmc_cpld_version is 'None' else "{}.{}".format(
-                int(mmc_cpld_version[2], 16), int(mmc_cpld_version[3], 16))
-
-            cpld_version["SMC"] = smc_cpld_version
-            cpld_version["MMC"] = mmc_cpld_version
-
-            component_versions["CPLD"] = cpld_version
-
-        component_versions["BIOS"] = bios_version.strip()
-        return str(component_versions)
